chore: unit-test-runner 3.0.0-rc.1 setup and example
