+++ demo-angular/nativescript.config.ts
@@ -7,5 +7,8 @@ export default {
   android: {
     v8Flags: '--expose_gc',
     markingMode: 'none'
+  },
+  cli: {
+    packageManager: 'npm'
   }
 } as NativeScriptConfig;
